refactor: 게시글 수집용 스레드 세이프 우선 순위 큐를 utill로 분리
